Move the patches list to a tab
